Moved code around to better write integration tests
